poc/pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Prova de conceito inicial da arquitetura multiagente para RAG híbrido"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "docling>=2.63.0",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.3",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
